+++ tests/integration-tests/framework/tests_configuration/config_renderer.py
@@ -299,7 +299,16 @@ def _check_or_create_capacity_reservations(config_file, os_parameters, instance_
                 )
                 end_date = datetime.now(timezone.utc) + timedelta(hours=hours)
                 specs.append((instance_type, os_platform, count, end_date, enable_placement_group))
-            candidate_regions = ["us-east-1", "us-east-2", "us-west-2", "eu-west-1"]
+            candidate_regions = [
+                "us-east-1",
+                "us-east-2",
+                "us-west-2",
+                "eu-west-1",
+                "ap-northeast-2",
+                "ap-southeast-2",
+                "ap-northeast-1",
+                "eu-west-2",
+            ]
             if len(specs) == 1:
                 # Single instance type reservation: check for existing reservation to be frugal.
                 # It is hard to implement such logic for multiple instance types reservation.
@@ -355,6 +364,12 @@ def _create_capacity_reservations(az_for_cr, regions, specs, var):  # noqa C901
     return False
 
 
+def _replace_last(string, old, new):
+    """Replaces the last occurrence of a substring in a string."""
+    parts = string.rsplit(old, 1)
+    return new.join(parts)
+
+
 def _resolve_instance_type_and_os(instance_type, instance_type_parameters, os, os_parameters):
     if "INSTANCE_TYPE" in instance_type:
         # The value of the Jinja INSTANCE_TYPE variable can contain a size or not, e.g. trn1.32xlarge vs trn1.
@@ -368,7 +383,8 @@ def _resolve_instance_type_and_os(instance_type, instance_type_parameters, os, o
         else:
             instance_type = instance_type_parameters.get(instance_type)
     else:
-        instance_type = instance_type.replace("_", ".")
+        instance_type = _replace_last(instance_type, "_", ".")
+        instance_type = instance_type.replace("_", "-")
     os_platform = "Linux/UNIX"
     if os is not None:
         if "OS" in os: